Moved intercom from JOIN page to CONTACT page
