+++ src/templates/join/index.js
@@ -1,9 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Link from "gatsby-link";
-import v3 from "uuid/v3";
 import withSegmentTracking from "../../utils/withSegmentTracking";
-import withIntercom from "../../utils/withIntercom";
 import TopJumbotron from "../../components/TopJumbotron";
 import "./styles.scss";
 
@@ -24,37 +22,7 @@ class JoinForm extends React.Component {
   }
 
   handleSubmit = () => {
-    const lead = this.getLead();
     this.props.trackGroup("Join", this.state.Company, this.state);
-    this.props.convertVisitorToLead();
-    this.props.updateLead(lead);
-  };
-
-  getLead = () => {
-    const {
-      FirstName,
-      LastName,
-      AgencyName,
-      Email: email,
-      ContactNumber: phone,
-    } = this.state;
-    const user_id = process.env.INTERCOM_ACCESS_TOKEN;
-    const company_id = v3(AgencyName, v3.DNS);
-    return {
-      user_id,
-      phone,
-      email,
-      name: `${FirstName} ${LastName}`,
-      companies: [
-        {
-          company_id,
-          name: AgencyName,
-          custom_attributes: {
-            "Primary Product": "LPMA",
-          },
-        },
-      ],
-    };
   };
 
   handleChange = ({target: {name, value}}) => {
@@ -129,7 +97,7 @@ JoinForm.propTypes = {
   trackGroup: PropTypes.func.isRequired,
 };
 
-const JoinFormWithSegmentTracking = withSegmentTracking(withIntercom(JoinForm));
+const JoinFormWithSegmentTracking = withSegmentTracking(JoinForm);
 
 class JoinPage extends React.PureComponent {
   render() {
